# StablingApi/Models/AbonementUsage.cs
namespace StablingApi.Models
{
    public class AbonementUsage
    {
        public int AbonementUsageId { get; set; }
        public int AbonementId { get; set; }
        public int TrainingId { get; set; }
    }
}