# StablingApi/swagger.json
{
...
  "swagger": "2.0",
  "info": {
    "title": "StablingApi",
    "description": "API ИС поддержки управления КСК",
    "version": "v1"
  },
  "paths": {
...
    "/api/Abonements/{abonementId}": {
      "get": {
        "tags": [
          "Abonements"
        ],
        "operationId": "Abonements_Get",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "abonementId",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/Abonement"
            }
          }
        }
      }
    },
    "/api/Abonements/ForShow/{abonementId}": {
      "get": {
        "tags": [
          "Abonements"
        ],
        "operationId": "Abonements_GetForShow",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "abonementId",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/AbonementForShow"
            }
          }
        }
      }
    },
    "/api/Abonements/ForShow/ByClient/{clientId}": {
      "get": {
        "tags": [
          "Abonements"
        ],
        "operationId": "Abonements_GetForShowByClient",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "clientId",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/AbonementForShow"
              }
            }
          }
        }
      }
    },
    "/api/Abonements/ByClient/{clientId}": {
      "get": {
        "tags": [
          "Abonements"
        ],
        "operationId": "Abonements_GetByClient",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "clientId",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Abonement"
              }
            }
          }
        }
      }
    },
    "/api/Abonements/ByAvailability/ForShow/{isAvailable}": {
      "get": {
        "tags": [
          "Abonements"
        ],
        "operationId": "Abonements_GetForShowByAvailability",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "type": "boolean",
            "name": "isAvailable",
            "in": "path",
            "required": true,
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/AbonementForShow"
              }
            }
          }
        }
      }
    },
    "/api/Abonements": {
      "post": {
        "tags": [
          "Abonements"
        ],
        "operationId": "Abonements_Create",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "name": "abonement",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Abonement"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "201": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/Abonement"
            }
          }
        }
      }
    },
    "/api/Abonements/Usage": {
      "post": {
        "tags": [
          "Abonements"
        ],
        "operationId": "Abonements_CreateUsage",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "name": "usage",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AbonementUsage"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "201": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/AbonementUsage"
            }
          }
        }
      }
    },
    "/api/Abonements/Usage/{abonementId}/{trainingId}": {
      "post": {
        "tags": [
          "Abonements"
        ],
        "operationId": "Abonements_CreateUsage2",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "abonementId",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          },
          {
            "type": "integer",
            "name": "trainingId",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "201": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/AbonementUsage"
            }
          }
        }
      }
    },
    "/api/Abonements/ChangeAvailability/{abonementId}": {
      "put": {
        "tags": [
          "Abonements"
        ],
        "operationId": "Abonements_ChangeAvailability",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "abonementId",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        }
      }
    },
    "/api/Abonements/PaidStatus/{abonementId}": {
      "put": {
        "tags": [
          "Abonements"
        ],
        "operationId": "Abonements_ChangePaidStatus",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "abonementId",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        }
      }
    },
...
    "/api/Abonements/Types/{id}": {
      "get": {
        "tags": [
          "Abonements"
        ],
        "operationId": "Abonements_GetType",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/AbonementType"
            }
          }
        }
      }
    },
    "/api/Abonements/{id}": {
      "delete": {
        "tags": [
          "Abonements"
        ],
        "operationId": "Abonements_Delete",
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "AbonementForShow": {
      "type": "object",
      "required": [
        "abonementId",
        "abonementTypeId",
        "typeName",
        "clientId",
        "clientName",
        "trainerId",
        "trainerName",
        "maxUses",
        "usesCount",
        "isAvailable",
        "isPaid",
        "openDateTime",
        "abonementFunds",
        "abonementLabel"
      ],
      "properties": {
        "abonementId": {
          "type": "integer",
          "format": "int32"
        },
        "abonementTypeId": {
          "type": "integer",
          "format": "int32"
        },
        "typeName": {
          "type": "string"
        },
        "clientId": {
          "type": "integer",
          "format": "int32"
        },
        "clientName": {
          "type": "string"
        },
        "trainerId": {
          "type": "integer",
          "format": "int32"
        },
        "trainerName": {
          "type": "string"
        },
        "maxUses": {
          "type": "integer",
          "format": "int32"
        },
        "usesCount": {
          "type": "integer",
          "format": "int32"
        },
        "isAvailable": {
          "type": "boolean"
        },
        "isPaid": {
          "type": "boolean"
        },
        "openDateTime": {
          "type": "string",
          "format": "date-time"
        },
        "abonementFunds": {
          "type": "number",
          "format": "double"
        },
        "abonementLabel": {
          "type": "string"
        }
      }
    },
    "Abonement": {
      "type": "object",
      "required": [
        "abonementId",
        "clientId",
        "abonementTypeId",
        "trainerId",
        "usesCount",
        "openDateTime",
        "isAvailable",
        "isPaid"
      ],
      "properties": {
        "abonementId": {
          "type": "integer",
          "format": "int32"
        },
        "clientId": {
          "type": "integer",
          "format": "int32"
        },
        "abonementTypeId": {
          "type": "integer",
          "format": "int32"
        },
        "trainerId": {
          "type": "integer",
          "format": "int32"
        },
        "usesCount": {
          "type": "integer",
          "format": "int32"
        },
        "openDateTime": {
          "type": "string",
          "format": "date-time"
        },
        "isAvailable": {
          "type": "boolean"
        },
        "isPaid": {
          "type": "boolean"
        }
      }
    },
    "AbonementUsage": {
      "type": "object",
      "required": [
        "abonementUsageId",
        "abonementId",
        "trainingId"
      ],
      "properties": {
        "abonementUsageId": {
          "type": "integer",
          "format": "int32"
        },
        "abonementId": {
          "type": "integer",
          "format": "int32"
        },
        "trainingId": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "AbonementType": {
      "type": "object",
      "required": [
        "abonementTypeId",
        "typeName",
        "typeUsesCount",
        "typePrice",
        "isAvailable"
      ],
      "properties": {
        "abonementTypeId": {
          "type": "integer",
          "format": "int32"
        },
        "typeName": {
          "type": "string"
        },
        "typeUsesCount": {
          "type": "integer",
          "format": "int32"
        },
        "typePrice": {
          "type": "integer",
          "format": "int32"
        },
        "isAvailable": {
          "type": "boolean"
        }
      }
    },
...
  }
}